add: gateway manager
