+++ includes/REST/PaymentController.php
@@ -1,6 +1,8 @@
 <?php
 namespace WeDevs\WePOS\REST;
 
+use \WeDevs\WePOS\Gateways\Manager as GatewayManager;
+
 /**
 * Payment API Controller
 */
@@ -31,7 +33,7 @@ class PaymentController extends \WC_REST_Orders_Controller {
         register_rest_route( $this->namespace, '/' . $this->base . '/gateways', array(
             array(
                 'methods'  => \WP_REST_Server::READABLE,
-                'callback' => array( $this, 'get_avaible_gateways' ),
+                'callback' => array( $this, 'get_available_gateways' ),
                 'permission_callback' => '__return_true',
                 'args'     => $this->get_collection_params()
             ),
@@ -73,10 +75,10 @@ class PaymentController extends \WC_REST_Orders_Controller {
      *
      * @since 1.0.0
      *
-     * @return void
+     * @return \WP_Error|\WP_HTTP_Response|\WP_REST_Response
      */
-    public function get_avaible_gateways( $request ) {
-        $available_gateways = \WePOS::init()->available_gateway();
+    public function get_available_gateways( $request ) {
+        $available_gateways = GatewayManager::available_gateway();
         $gateways = [];
 
         foreach ( $available_gateways as $class => $path ) {
@@ -91,10 +93,10 @@ class PaymentController extends \WC_REST_Orders_Controller {
      *
      * @since 1.0.0
      *
-     * @return void
+     * @return \WP_Error|\WP_HTTP_Response|\WP_REST_Response
      */
     public function process_payment( $request ) {
-        $available_gateways = \WePOS::init()->available_gateway();
+        $available_gateways = GatewayManager::available_gateway();
         $chosen_gateway = '';
 
         if ( empty( $request['id'] ) ) {
